refactor: 뚜두 목록 검색 시 쿼리 이후 로직 모듈 위치 변경
