+++ application/application-common/src/main/java/com/ddudu/application/common/dto/ddudu/DduduCursorDto.java
@@ -1,0 +1,5 @@
+package com.ddudu.application.common.dto.ddudu;
+
+public record DduduCursorDto(String cursor, SimpleDduduSearchDto ddudu) {
+
+}
