+++ application/planning-application/src/main/java/com/ddudu/application/planning/ddudu/service/DduduSearchService.java
@@ -1,5 +1,6 @@
 package com.ddudu.application.planning.ddudu.service;
 
+import com.ddudu.application.common.dto.ddudu.DduduCursorDto;
 import com.ddudu.application.common.dto.ddudu.SimpleDduduSearchDto;
 import com.ddudu.application.common.dto.ddudu.request.DduduSearchRequest;
 import com.ddudu.application.common.dto.scroll.response.ScrollResponse;
@@ -9,6 +10,7 @@ import com.ddudu.application.common.port.user.out.UserLoaderPort;
 import com.ddudu.common.annotation.UseCase;
 import com.ddudu.common.exception.DduduErrorCode;
 import com.ddudu.domain.user.user.aggregate.User;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,13 +28,36 @@ public class DduduSearchService implements DduduSearchUseCase {
         loginId,
         DduduErrorCode.USER_NOT_EXISTING.getCodeName()
     );
-
-    return dduduSearchPort.search(
+    List<DduduCursorDto> ddudusWithCursor = dduduSearchPort.search(
         user.getId(),
         request.getScroll(),
         request.getQuery(),
         request.getIsMine()
     );
+
+    return getScrollResponse(ddudusWithCursor, request.getSize());
+  }
+
+  private ScrollResponse<SimpleDduduSearchDto> getScrollResponse(
+      List<DduduCursorDto> ddudusWithCursor,
+      int size
+  ) {
+    List<SimpleDduduSearchDto> simpleDdudus = ddudusWithCursor.stream()
+        .limit(size)
+        .map(DduduCursorDto::ddudu)
+        .toList();
+    String nextCursor = getNextCursor(ddudusWithCursor, size);
+
+    return ScrollResponse.from(simpleDdudus, nextCursor);
+  }
+
+  private String getNextCursor(List<DduduCursorDto> ddudusWithCursor, int size) {
+    if (ddudusWithCursor.size() > size) {
+      return ddudusWithCursor.get(size - 1)
+          .cursor();
+    }
+
+    return null;
   }
 
 }
