--- infra/planning-infra-mysql/src/main/java/com/ddudu/infra/mysql/planning/ddudu/dto/DduduCursorDto.java
@@ -1,7 +1,0 @@
-package com.ddudu.infra.mysql.planning.ddudu.dto;
-
-import com.ddudu.application.common.dto.ddudu.SimpleDduduSearchDto;
-
-public record DduduCursorDto(String cursor, SimpleDduduSearchDto ddudu) {
-
-}
